chore[Op#56922]: extract oauth access grantable concern
