+++ modules/storages/app/controllers/storages/admin/project_storages_controller.rb
@@ -27,6 +27,8 @@
 #++
 
 class Storages::Admin::ProjectStoragesController < Projects::SettingsController
+  include Storages::OAuthAccessGrantable
+
   model_object Storages::ProjectStorage
 
   before_action :find_model_object, only: %i[oauth_access_grant edit update destroy destroy_info]
@@ -70,7 +72,7 @@ class Storages::Admin::ProjectStoragesController < Projects::SettingsController
     end
   end
 
-  def oauth_access_grant # rubocop:disable Metrics/AbcSize
+  def oauth_access_grant
     @project_storage = @object
     storage = @project_storage.storage
     auth_state = ::Storages::Peripherals::StorageInteraction::Authentication
@@ -79,14 +81,10 @@ class Storages::Admin::ProjectStoragesController < Projects::SettingsController
     if auth_state == :connected
       redirect_to(external_file_storages_project_settings_project_storages_path)
     else
-      nonce = SecureRandom.uuid
-      cookies["oauth_state_#{nonce}"] = {
-        value: { href: external_file_storages_project_settings_project_storages_url(project_id: @project_storage.project_id),
-                 storageId: @project_storage.storage_id }.to_json,
-        expires: 1.hour
-      }
-      session[:oauth_callback_flash_modal] = storage_oauth_access_granted_modal(storage:)
-      redirect_to(storage.oauth_configuration.authorization_uri(state: nonce), allow_other_host: true)
+      open_redirect_to_storage_authorization_with(
+        callback_url: external_file_storages_project_settings_project_storages_url(project_id: @project_storage.project_id),
+        storage: @project_storage.storage
+      )
     end
   end
 
@@ -156,36 +154,17 @@ class Storages::Admin::ProjectStoragesController < Projects::SettingsController
   end
 
   def redirect_to_project_storages_path_with_oauth_access_grant_confirmation
-    if storage_oauth_access_granted?
+    if storage_oauth_access_granted?(storage: @project_storage.storage)
       redirect_to external_file_storages_project_settings_project_storages_path
     else
       redirect_to_project_storages_path_with_nudge_modal
     end
   end
 
-  def storage_oauth_access_granted?
-    OAuthClientToken
-      .exists?(user: current_user, oauth_client: @project_storage.storage.oauth_client)
-  end
-
   def redirect_to_project_storages_path_with_nudge_modal
     redirect_to(
       external_file_storages_project_settings_project_storages_path,
-      flash: { modal: oauth_access_grant_nudge_modal(project_storage: @project_storage) }
+      flash: { modal: project_storage_oauth_access_grant_nudge_modal(project_storage: @project_storage) }
     )
-  end
-
-  def oauth_access_grant_nudge_modal(project_storage:)
-    {
-      type: Storages::ProjectStorages::OAuthAccessGrantNudgeModalComponent.name,
-      parameters: { project_storage: project_storage.id }
-    }
-  end
-
-  def storage_oauth_access_granted_modal(storage:)
-    {
-      type: Storages::Admin::Storages::OAuthAccessGrantedModalComponent.name,
-      parameters: { storage: storage.id }
-    }
   end
 end
